所有在售列表(本平台+其他平台)

+++ yudao-module-steam/yudao-module-steam-biz/src/main/java/cn/iocoder/yudao/module/steam/dal/dataobject/otherselling/OtherSellingDO.java
@@ -122,7 +122,7 @@ public class OtherSellingDO extends BaseDO {
     /**
      * 出售用户id
      */
-    private Integer sellingUserId;
+    private Long sellingUserId;
     /**
      * 出售平台id
      */
